requirements.txt
pillow
sv-ttk
pySmartDL